arm64: DT: MSM8956: Declare wakegic as interrupt controller
Add the missing property "interrupt-controller" to wakegic node.
Without the property, the kernel reports the following error:
[    0.000000] OF: of_irq_init: children remain, but no parents
It subsequently skips the wakegic node, which causes mpm to not initialise.
Without having mpm initialised, sys_pm_ops is not registered in the lpm driver.
That then leads the lpm driver to skip over any lpm level that notifies rpm.
Avoid this chain of events by correcting the wakegic node declaration.

Change-Id: I69613170d2a4a6ae45470f119c70b548c763ae8e